+++ sdk/storage/Azure.Storage.DataMovement/CHANGELOG.md
@@ -58,6 +58,10 @@
 - [BREAKING CHANGE] Changed `StorageTransferProgress` constructor from `public` to `protected internal`. 
 - [BREAKING CHANGE] Renamed `StorageTransferProgress` to `DataTransferProgress`.
 - [BREAKING CHANGE] Renamed `StorageTransferEventArgs` to `DataTransferEventArgs`.
+- [BREAKING CHANGE] Removed `position` parameter from `StorageResourceSingle.WriteFromStreamAsync`. Use `StorageResourceWriteToOffsetOptions.Position` instead.
+- [BREAKING CHANGE] Made parameter `completeLength` from `StorageResourceSingle.CopyBlockFromUriAsync` mandatory.
+- [BREAKING CHANGE] Moved `DataTransferOptions.ProgressHandler` to `DataTransferOptions.ProgressHandlerOptions`.
+- [BREAKING CHANGE] Removed default constructor for `ProgressHandlerOptions`. Use `ProgressHandlerOptions(IProgress<DataTransferProgress>, bool)` instead.
 
 ### Bugs Fixed